tryg-medborger/exp/requirements.txt
transformers
datasets
numpy>=1.16.6
torch_optimizer